modify reviewlist into review card
